tentatives
