+++ src/app/autocomp-input/autocomp-input.component.html
@@ -1,11 +1,13 @@
-<form class="example-form">
-  <mat-form-field class="example-full-width">
-    <input name="inputCity" type="text" placeholder="Pick one" matInput [(ngModel)]="valueInput"
-      [matAutocomplete]="auto">
-    <mat-autocomplete #auto="matAutocomplete">
-      <mat-option *ngFor="let option of options" [value]="option.nom">
-        {{option.nom}}
-      </mat-option>
-    </mat-autocomplete>
-  </mat-form-field>
-</form>
+
+<mat-form-field class="example-full-width">
+	<input name="inputCity" type="text" [placeholder]="placeHolder" matInput [(ngModel)]="valueInput"
+	[matAutocomplete]="auto">
+	<mat-error *ngIf="cityCtrl.hasError('required')">
+		Ville <strong>obligatoire</strong>
+	</mat-error>
+	<mat-autocomplete #auto="matAutocomplete">
+		<mat-option *ngFor="let option of options" [value]="option.nom">
+			{{option.nom}}
+		</mat-option>
+	</mat-autocomplete>
+</mat-form-field>
