+++ src/app/register/register.component.css
@@ -1,0 +1,70 @@
+.login-container {
+    padding-top: 50px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.login-card {
+    width: 680px;
+    border-radius: 14px;
+    background-color: #e7e6e6;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    padding-bottom: 20px;
+}
+
+.login-card__content{
+    width: 60%;
+    margin: 20px auto 0 auto;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+}
+
+.login-card__title {
+    font-family: Roboto;
+    font-size: 20px;
+    font-weight: 500;
+    font-style: normal;
+    font-stretch: normal;
+    line-height: 1.4;
+    letter-spacing: normal;
+    color: rgba(0, 0, 0, 0.54);
+}
+
+.login-card__form{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    color: rgba(0, 0, 0, 0.54);
+}
+
+.login-card__footer{
+    width: 100%;
+    margin-top: 30px;
+    display: flex;
+    justify-content: flex-end;
+}
+
+.login-card__button{
+    margin-right: 10px;
+}
+
+.login-card__button--cancel{
+    background-color: #4a4a4a;
+    color: white;
+}
+
+.login-card__button--validate{
+    background-color: #17998c;
+    color: white;
+}
+
+.login-card__error{
+    border-radius: 10px;
+    background-color: #be1d1d;
+    color: white;
+    padding: 10px;
+}
