+++ src/app/register/register.component.html
@@ -1,51 +1,51 @@
-
 <app-header-picture></app-header-picture>
+<div class="container">
+	<div class="login-container">
+		<div class="login-card">
+			<h2 class="login-card__title">Formulaire d'inscription</h2>
+			<div class="login-card__content">
+				<form (ngSubmit)="handleSubmit()" [formGroup]="registerForm">
+					<mat-form-field class="login-card__input">
+						<input matInput placeholder="Identifiant" formControlName="username">
+						<mat-error *ngIf="usernameCtrl.hasError('required')">
+							Identifiant <strong>obligatoire</strong>
+						</mat-error>
+					</mat-form-field >
+					<div formGroupName="passwordGroup">
+						<mat-form-field class="login-card__input">
+							<input matInput placeholder="Mot de passe" formControlName="password" [errorStateMatcher]="passwordMatcher">
+							<mat-error *ngIf = "passwordCtrl.hasError('pattern') && !passwordCtrl.hasError('required')" >
+								Le mot de passe doit contenir au moins <strong>8 caractères</strong> dont <strong>une minuscule</strong>, <strong>une majuscule</strong> et <strong>un chiffre</strong>.
+							</mat-error>
+							<mat-error *ngIf="passwordCtrl.hasError('required')">
+								Mot de passe <strong>obligatoire</strong>
+							</mat-error>
+						</mat-form-field>
+						<mat-form-field class="login-card__input">
+							<input matInput placeholder="Confirmez le mot de passe" formControlName="confirmPassword" [errorStateMatcher]="passwordMatcher">
+							<mat-error *ngIf="passwordGroup.hasError('notSame')">
+								Mot de passe incorrect
+							</mat-error>
+						</mat-form-field>
+					</div>
+					<mat-form-field class="login-card__input">
+						<input matInput placeholder="Email" formControlName="email" [errorStateMatcher]="matcher">
+						<mat-error *ngIf="emailCtrl.hasError('required')">
+							Email <strong>obligatoire</strong>
+						</mat-error>
+						<mat-error *ngIf="emailCtrl.hasError('email')">
+							Email <strong>invalide</strong>
+						</mat-error>
+						
+					</mat-form-field>
+					<app-autocomp-input class="login-card__input"  autocompType="communes" placeHolder="Ville" [cityCtrl]="cityCtrl" (nameSelect)="handleCitySelect($event)" (click)="outputCityCtrl()">
+					</app-autocomp-input>
+					<mat-checkbox>Êtes vous un artiste proposant des prestations privées?</mat-checkbox>
 
-<mat-card>
-	<mat-card-title>Formulaire d'inscription</mat-card-title>
-	<form (ngSubmit)="handleSubmit()" [formGroup]="registerForm">
-		<mat-form-field>
-			<input matInput placeholder="Identifiant" formControlName="username">
-			<mat-error *ngIf="usernameCtrl.hasError('required')">
-					Identifiant <strong>obligatoire</strong>
-				</mat-error>
-		</mat-form-field>
-		<div formGroupName="passwordGroup">
-			<mat-form-field >
-				<input matInput placeholder="Mot de passe" formControlName="password" [errorStateMatcher]="passwordMatcher">
-				<mat-error *ngIf = "passwordCtrl.hasError('pattern') && !passwordCtrl.hasError('required')" >
-					Le mot de passe doit contenir au moins <strong>8 caractères</strong> dont <strong>une minuscule</strong>, <strong>une majuscule</strong> et <strong>un chiffre</strong>.
-				</mat-error>
-				<mat-error *ngIf="passwordCtrl.hasError('required')">
-					Mot de passe <strong>obligatoire</strong>
-				</mat-error>
-			</mat-form-field>
-			<mat-form-field >
-				<input matInput placeholder="Confirmez le mot de passe" formControlName="confirmPassword" [errorStateMatcher]="passwordMatcher">
-				<mat-error *ngIf="passwordGroup.hasError('notSame')">
-					Mot de passe incorrect
-				</mat-error>
-			</mat-form-field>
+					<button mat-raised-button type="submit" [disabled]="registerForm.invalid" >Valider</button>
+					<button mat-raised-button (click)="testNavigate()">Annuler</button>
+				</form>
+			</div>
 		</div>
-		<mat-form-field>
-			<input matInput placeholder="Email" formControlName="email" [errorStateMatcher]="matcher">
-			<mat-error *ngIf="emailCtrl.hasError('email')">
-					Email <strong>invalide</strong>
-				</mat-error>
-			<mat-error *ngIf="emailCtrl.hasError('required')">
-					Email <strong>obligatoire</strong>
-				</mat-error>
-		</mat-form-field>
-		<mat-form-field>
-			<input matInput placeholder="Ville" formControlName="city" >
-			<mat-error *ngIf="cityCtrl.hasError('required')">
-					Ville <strong>obligatoire</strong>
-				</mat-error>
-		</mat-form-field>
-		<mat-checkbox>Êtes vous un artiste proposant des prestations privées?</mat-checkbox>
-
-		<button mat-raised-button type="submit" [disabled]="registerForm.invalid" >Valider</button>
-		
-	</form>
-	<button mat-raised-button (click)="testNavigate()">Annuler</button>
-</mat-card>
+	</div>
+</div>
